+++ wasm/src/quadtree.rs
@@ -30,18 +30,18 @@ impl Quadtree {
         }
     }
 
-    pub fn inside<B: Walkable<i32, 2>>(bbox: &B) -> Quadtree {
+    pub fn inside<B: Walkable<i64, 2>>(bbox: &B) -> Quadtree {
         Quadtree {
             root: Node::new(Area::surrounding(bbox))
         }
     }
 
     #[inline]
-    pub fn has(&self, point: &Point2<i32>) -> bool {
+    pub fn has(&self, point: &Point2<i64>) -> bool {
         self.root.area.holds(point) && self.root.has(point)
     }
 
-    pub fn query<R: Holds<Point2<i32>> + Overlaps<Area>, B: Intersection<Area, Output=R>>(&self, bbox_a: B) -> Query<R> {
+    pub fn query<R: Holds<Point2<i64>> + Overlaps<Area>, B: Intersection<Area, Output=R>>(&self, bbox_a: B) -> Query<R> {
         let bbox = bbox_a.intersection(&self.root.area);
         Query::new(bbox, &self.root)
     }
@@ -52,7 +52,7 @@ impl Quadtree {
     }
 
     #[inline]
-    pub fn insert(&mut self, point: Point2<i32>) {
+    pub fn insert(&mut self, point: Point2<i64>) {
         if self.root.area.holds(&point) {
             self.root.insert(Tree::Leaf(point), &point);
         } else {
@@ -61,7 +61,7 @@ impl Quadtree {
     }
 
     #[inline]
-    pub fn remove(&mut self, point: &Point2<i32>) {
+    pub fn remove(&mut self, point: &Point2<i64>) {
         if self.root.area.holds(point) {
             self.root.remove(point);
         }
@@ -77,7 +77,7 @@ impl Default for Quadtree {
 }
 
 impl<'a> IntoIterator for &'a Quadtree {
-    type Item = &'a Point2<i32>;
+    type Item = &'a Point2<i64>;
     type IntoIter = Iter<'a>;
 
     #[inline]
