+++ wasm/src/quadtree/area.rs
@@ -9,8 +9,8 @@ use crate::traits::overlap::Overlaps;
 /// An area in the quadtree
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 pub struct Area {
-    pub anchor: Point2<i32>,
-    pub size: u32,
+    pub anchor: Point2<i64>,
+    pub size: u64,
 }
 
 impl Area {
@@ -18,13 +18,13 @@ impl Area {
     pub fn global() -> Area {
         Area {
             anchor: Point2::origin(),
-            size: u32::MAX,
+            size: u64::MAX,
         }
     }
 
-    fn search_area(start: &Point2<i32>, end: &Point2<i32>, mut bits: u32) -> Area {
-        while bits < u32::BITS {
-            let mask = (u32::MAX << bits) as i32;
+    fn search_area(start: &Point2<i64>, end: &Point2<i64>, mut bits: u64) -> Area {
+        while bits < u64::BITS {
+            let mask = (u64::MAX << bits) as i64;
 
             let area = Area {
                 anchor: point![start.x & mask, start.y & mask],
@@ -43,7 +43,7 @@ impl Area {
     }
 
     /// Returns area surrounding given bbox
-    pub fn surrounding<B: Walkable<i32, 2>>(bbox: &B) -> Area {
+    pub fn surrounding<B: Walkable<i64, 2>>(bbox: &B) -> Area {
         let start = bbox.first_point();
         let end = bbox.last_point();
 
@@ -51,7 +51,7 @@ impl Area {
             (Some(start), Some(end)) => {
                 if quarter(&Point2::origin(), start) == quarter(&Point2::origin(), end) {
                     let size = end - start;
-                    let bits = u32::BITS - max(size.x.unsigned_abs(), size.y.unsigned_abs()).leading_zeros();
+                    let bits = u64::BITS - max(size.x.unsigned_abs(), size.y.unsigned_abs()).leading_zeros();
 
                     Area::search_area(start, end, bits)
                 } else {
@@ -88,11 +88,11 @@ impl Area {
         Area::search_area(&start, &end, bits)
     }
 
-    fn center(&self) -> Point2<i32> {
-        if self.size == u32::MAX {
+    fn center(&self) -> Point2<i64> {
+        if self.size == u64::MAX {
             Point2::origin()
         } else {
-            let delta = self.size as i32 / 2;
+            let delta = self.size as i64 / 2;
 
             point![
                 self.anchor.x + delta,
@@ -102,7 +102,7 @@ impl Area {
     }
 
     #[inline]
-    pub fn quarter(&self, point: &Point2<i32>) -> Quarter {
+    pub fn quarter(&self, point: &Point2<i64>) -> Quarter {
         quarter(&self.center(), point)
     }
 }
@@ -110,19 +110,19 @@ impl Area {
 // Utils
 impl Division for Area {
     #[inline]
-    fn anchor(&self) -> &Point2<i32> {
+    fn anchor(&self) -> &Point2<i64> {
         &self.anchor
     }
 
     #[inline]
-    fn size(&self) -> u32 {
+    fn size(&self) -> u64 {
         self.size
     }
 }
 
 impl<D: Division> Holds<D> for Area {
     fn holds(&self, object: &D) -> bool {
-        if self.size == u32::MAX {
+        if self.size == u64::MAX {
             true
         } else if object.size() > self.size {
             false
@@ -131,16 +131,16 @@ impl<D: Division> Holds<D> for Area {
 
             self.anchor.iter()
                 .zip(object.anchor().iter())
-                .all(|(a, o)| (a ^ o) as u32 <= left)
+                .all(|(a, o)| (a ^ o) as u64 <= left)
         }
     }
 }
 
-impl Intersection<Area> for Range<Point2<i32>> {
-    type Output = Range<Point2<i32>>;
+impl Intersection<Area> for Range<Point2<i64>> {
+    type Output = Range<Point2<i64>>;
 
     fn intersection(&self, lhs: &Area) -> Self::Output {
-        if lhs.size == u32::MAX {
+        if lhs.size == u64::MAX {
             self.clone()
         } else {
             self.intersection(&Range::from(lhs))
@@ -148,11 +148,11 @@ impl Intersection<Area> for Range<Point2<i32>> {
     }
 }
 
-impl Intersection<Area> for RangeInclusive<Point2<i32>> {
-    type Output = BBox<i32, 2>;
+impl Intersection<Area> for RangeInclusive<Point2<i64>> {
+    type Output = BBox<i64, 2>;
 
     fn intersection(&self, lhs: &Area) -> Self::Output {
-        if lhs.size == u32::MAX {
+        if lhs.size == u64::MAX {
             BBox::from(self.clone())
         } else {
             self.intersection(&Range::from(lhs))
@@ -160,12 +160,12 @@ impl Intersection<Area> for RangeInclusive<Point2<i32>> {
     }
 }
 
-impl Overlaps<Area> for Range<Point2<i32>> {
+impl Overlaps<Area> for Range<Point2<i64>> {
     fn overlap(&self, lhs: &Area) -> bool {
-        if lhs.size == u32::MAX {
+        if lhs.size == u64::MAX {
             true
         } else {
-            let size = lhs.size as i32;
+            let size = lhs.size as i64;
 
             self.start.x < (lhs.anchor.x + size) && self.start.y < (lhs.anchor.y + size) && self.end.x >= lhs.anchor.x && self.end.y >= lhs.anchor.y
         }
@@ -173,9 +173,9 @@ impl Overlaps<Area> for Range<Point2<i32>> {
 }
 
 // Conversions
-impl From<&Area> for Range<Point2<i32>> {
+impl From<&Area> for Range<Point2<i64>> {
     fn from(value: &Area) -> Self {
-        value.anchor..Point2::new(value.anchor.x + value.size as i32, value.anchor.y + value.size as i32)
+        value.anchor..Point2::new(value.anchor.x + value.size as i64, value.anchor.y + value.size as i64)
     }
 }
