+++ wasm/src/quadtree/division.rs
@@ -2,6 +2,6 @@ use na::Point2;
 
 /// Represents a quadtree division
 pub trait Division {
-    fn anchor(&self) -> &Point2<i32>;
-    fn size(&self) -> u32;
+    fn anchor(&self) -> &Point2<i64>;
+    fn size(&self) -> u64;
 }
